pyproject.toml
[project]
name = "biollm"
version = "0.1.0"
description = ""
authors = [
    {name = "HowardKuo2025",email = "howard.futurenest@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic (>=2.12.5,<3.0.0)",
    "google-genai (>=1.56.0,<2.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "lxml (>=6.0.2,<7.0.0)",
    "curl-cffi (>=0.14.0,<0.15.0)",
    "pypdf (>=6.5.0,<7.0.0)",
    "elsapy (>=0.5.1,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
